Pull out default rails actions into a spec file
Moved GET new, POST create, DELETE destroy, GET show, GET index. Changed my mind and put back GET show_comments, GET hide_comments, GET cancel_comment_delete, and GET cancel_comment_edit after initially moving them as well, that's why the blame is hard to follow.
